backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.0
Werkzeug==2.0.1
Flask-SocketIO==5.1.1
python-dateutil==2.8.2
eventlet==0.30.2 